Remove logging fix to seperate PR. Relocate conversion script to avoid logging circular import issue
