+++ deepspeed/inference/engine.py
@@ -7,7 +7,7 @@ import torch
 import time
 import os
 from deepspeed import comm as dist
-from deepspeed.utils import log_dist
+from deepspeed.utils.logging import log_dist
 
 from torch.nn.modules import Module
 from packaging import version as pkg_version
