+++ deepspeed/ops/adam/cpu_adam.py
@@ -5,7 +5,8 @@
 
 import torch
 from cpuinfo import get_cpu_info
-from deepspeed.utils import logger, should_log_le
+from deepspeed.utils import logger
+from deepspeed.utils.logging import should_log_le
 from deepspeed.ops.op_builder import CPUAdamBuilder
