+++ deepspeed/ops/lion/cpu_lion.py
@@ -6,7 +6,7 @@
 import torch
 from cpuinfo import get_cpu_info
 from deepspeed.utils import logger
-from deepspeed.utils import should_log_le
+from deepspeed.utils.logging import should_log_le
 from deepspeed.ops.op_builder import CPULionBuilder
