+++ deepspeed/ops/transformer/inference/diffusers_attention.py
@@ -8,7 +8,7 @@ import torch
 from torch.autograd import Function
 import torch.nn as nn
 from packaging import version as pkg_version
-from deepspeed.utils import log_dist
+from deepspeed.utils.logging import log_dist
 from deepspeed.accelerator import get_accelerator
 from deepspeed.ops.op_builder import InferenceBuilder
