+++ deepspeed/runtime/swap_tensor/partitioned_optimizer_swapper.py
@@ -8,7 +8,7 @@ Functionality of swapping optimizer tensors to/from (NVMe) storage devices.
 
 import torch
 
-from deepspeed.utils import logger
+from deepspeed.utils.logging import logger
 from deepspeed.ops.op_builder import AsyncIOBuilder
 from deepspeed import comm as dist
