+++ deepspeed/runtime/swap_tensor/utils.py
@@ -7,7 +7,7 @@ Functionality of swapping tensors to/from (NVMe) storage devices.
 """
 
 import torch
-from deepspeed.utils import logger
+from deepspeed.utils.logging import logger
 from deepspeed.accelerator import get_accelerator
 
 from deepspeed import comm as dist
